ai/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
aiomysql==0.2.0

# AI/ML
google-generativeai==0.3.2
sentence-transformers==2.2.2
chromadb==0.4.18

# MCP
fastmcp==0.1.0

# Monitoring
ddtrace==1.19.0

# Utilities
python-dotenv==1.0.0
jinja2==3.1.2